[test] Re-run failing Lit tests several times to catch non-determinism
If a Lit test fails, re-run it 5 times to increase our odds of
detecting non-determinism early. If the test turns out to
be flaky, it will fail and be reported as a FLAKYFAIL to stress the
encountered non-determinism.

Long tests and tests in `test/Interop/Cxx` are deliberately exempted
from retries to avoid timeouts and unreasonable spikes in turnaround
time.