Add isolation info to ExtInfo for SubscriptDecls
Subscripts with `isolated` arguments currently do not pass a compiler
assertion about the consistency of the function type w.r.t. isolation.
This is because the `ExtInfo` for `SubscriptDecl`s doesn't currently
have isolation information applied (though this works in practice);
annotating the `ExtInfo` with isolation allows this assertion to pass.