[cxx-interop] Avoid cycles when importing certain enums
This fixes an assertion failure:
```
Assertion failed: (builtinInitializer), function setBuiltinInitializer at Expr.cpp:1092
```

This was happening because of a cycle: ClangImporter tried to import an enum, which triggered a conformance lookup for `Int`, which triggered Clang module loading of the enum's owning module.

The proper fix might be avoiding this kind of cycles, but that would be a much more involved change.

rdar://163596214