[Swiftify] update ConstantCount.swift test (NFC)
Removes -ignore-availability so that we get all the diagnostics.
This requires the functions to be made public, since the overloads are
_alwaysEmitIntoClient.

Renames the noescape* overloads lacking .nonescaping, and verifies that
their overloads are actually emitted. The rename did not cause them to
be emitted, the name was just misleading.

This test was already passing, but now it has better coverage.

rdar://156418990