[cxx-interop] Suppress inaccessible-base warning from test
Clang is helpfully telling us that one of the base types is
inaccessible, except we don't really care about that because we are
only using this type for testing other diagnostics. That warning is
getting caught by the newly comprehensive swiftc diagnostic verifier,
so we suppress it with a Clang pragma.

rdar://163875420