[Custom availability] Add some hacky serialization
Custom availability serialization is based on serializing the
underlying declaration that describes the availability domain. Those
don't exist yet, so conjure up a VarDecl that we use just for the
name. When deserializing based on one of these declarations, do a
lookup into the module. This is quite gross, but lets us do more
experimentation.