Introduce a custom availability domain for Unicode
Tag the C functions used by the standard library for Unicode
normalization as being part of a new availability domain called
"Unicode". Mark anything else that (transitively) depends
on those functions as requiring Unicode, so that we get
the full surface area of the standard library that somehow depends on
the Unicode data tables (for normalization).

By default, this availability domain is always available. One can
override this default to disable the availability domain (all uses of
these APIs will be rejected) or enable it as conditional (all uses of
the APIs must be marked with `@available(Unicode)`). The
standard library, in Embedded Swift mode, builds with this domain
being conditional.

For Embedded Swift, this provides a better diagnostic experience when
one wants to avoid using the Unicode normalization tables. The
command-line interface itself is currently a mess, but in the future
we can provide a driver-level flag to either link in the Unicode
tables (via the swiftUnicodeDataTables library) or disable this
availability domain.

Part of rdar://151212706.

FIXUP test for UnicodeNormalization availability