[ClangImporter] Do not import enum when already imported via DeclContext
Before we import an enum, we import its DeclContext. If that DeclContext
gets resolved to a struct that contains that field of that enum type,
e.g.:

    typedef CF_OPTIONS(uint32_t, MyFlags) {
      ...
    } CF_SWIFT_NAME(MyCtx.Flags);

    struct MyStruct {
      MyFlags flags;
      ...
    } CF_SWIFT_NAME(MyCtx);

We end up with a cycle, i.e., MyFlags -> MyCtx -> MyStruct -> MyFlags.

Existing cycle-breaking mechanisms seem to prevent this from looping
infinitely, but it leads to us importing two copies of the enum, which
can cause errors later during CodeGen.

This patch breaks the cycle by checking the cache again.

N.B. as of this commit, the circular-import-as-member.swift test doesn't
actually fail *without* the fix, because ClangImporter is quite
lenient about importing the same clang::Decl multiple times. An earlier
attempt at this patch ABORT()'d the compiler as soon as a duplicate is
detected, but that led to more isseus beyond the scope of this patch.

rdar://162317760