[cxx interop] Broaden support for the 'swift_private' attribute.
Motivation: It is often desirable to explicitly hide C++ declaraitons from
swift either because some entry-points are known to be unsafe to call from swift
or to suppress "unavailable (cannot import)" messages. Before this change, it
was possible to exclude declarationis by using '#if !defined(__swift__)`, but
that approach is often unsafe because it can lead to undetected ODR violations
(undetected because linkers do not check ODR by default), which can lead to
memory corruption. A better approach is to use an annotation that excludes
declarations from being visible to swift without affecting class definitions to
avoid divergent v-tables and object layouts.  `__attribute__((swift_private))`
does this, but that attribute was not being checked on most types of C++
declarations.

Description of changes:
* Added a check for the swift_private attribute in
  `ClangImporter::Implementation::importDeclImpl` in order to exclude
  declarations from import.
* Added test coverage for use of the swift_private attribute on methods,
  fields, functions and classes in test ClangImporter/cxx_interop.swift