[cxx-interop] Implicitly defined copy and move constructors
If a C++ type doesn't have a user provided copy constructor, Clang will declare one implicitly, unless one of its fields is not copy-constructible.
Clang assumes that any std::vector is always copy-constructible, including a std::vector of a non-copy constructible type.
So, let's use the CxxValueSemantics implemented in #84340 to figure out if we should ask Clang to define the copy constructor for such types.

This also includes types that have a defaulted, user-provided copy constructor.