SILGen: fix cast of Optional<any Sendable> -> Optional<Any>
In some cases, at least with ObjC bridging, we could
end up throwing `SGF.emitSubstToOrigValue` into an infinite
recursion through `Transform::transform`, in the case of
the optional-to-optional conversion:

  Optional<any Sendable> -> Optional<Any>

There's special handling for such conversions in SILGen,
and the problem is its old recognition of `any Sendable` being
"equivalent" in terms of ABI to `Any` relied on type equality.

While the types are not equal at compile time, they are at
runtime. Plus, they are ABI equivalent.

resolves rdar://163872193